seperate components for expost options (css, link, svg)
